fix remote DoS via special nicks
I nickname this the Lalufu exploit:

Tweets by special nicknames like '__indexes', '__windows' and so on
will corrupt internal data structures and most likely result in a "Not
an ARRAY reference" error and crash the whole twirssi.pl script.
Subsequent restarts of the script (if the crash is noticed in the
first place) will most likely result in another crash as the offending
tweet will most likely be downloaded again.

It is trivial to register one of these usernames and with a single
tweet containing the #twirssi hashtag, it is possible to crash many
twirssi instances netwide at once as twirssi users are likely to
follow the #twirssi hashtag.

The fix is rather simple and blunt: it blackholes any tweets coming
from those nicks.

In the long run, the internal $state reprensentation should be
refactored.